requirements.txt
Django==4.0.6
numpy==1.20.3
opencv_python==4.6.0.66
Pillow==9.2.0
tensorflow==2.7.0